Finished Appointment lsit and create views

+++ service/api/service_rest/views.py
@@ -13,6 +13,7 @@ class AutomobileVOEncoder(ModelEncoder):
 class TechnicianListEncoder(ModelEncoder):
     model = Technician
     properties = [
+        "id",
         "first_name",
         "last_name",
         "employee_id",
@@ -26,7 +27,6 @@ class AppointmentListEncoder(ModelEncoder):
         "date_time",
         "reason",
         "status",
-        "technician",
     ]
 
 
@@ -40,18 +40,33 @@ class AppointmentDetailEncoder(ModelEncoder):
         "status",
         "technician",
     ]
+    encoders = {
+        "technician": TechnicianListEncoder(),
+    }
 
 
 @require_http_methods(["GET", "POST"])
-def technician_list(request):
+def technician_list(request, id=None):
     if request.method == "GET":
-        technicians = Technician.objects.all()
+        if id is None:
+            technicians = Technician.objects.all()
+        else:
+            technicians = Technician.objects.filter(id=id)
         return JsonResponse(
             {"technicians": technicians},
             encoder=TechnicianListEncoder
         )
     else:
         content = json.loads(request.body)
+        try:
+            tech = Technician.objects.get(id=content["technician"])
+            content["technician"] = tech
+        except Technician.DoesNotExist:
+            return JsonResponse(
+                {"message": "Invalid technician id"},
+                status=400,
+            )
+
         technician = Technician.objects.create(**content)
         return JsonResponse(
             technician,
@@ -59,17 +74,52 @@ def technician_list(request):
             safe=False
         )
 
-@require_http_methods(["DELETE"])
+@require_http_methods(["DELETE", "GET"])
 def delete_technician(request, id):
-    try:
+    if request.method == "DELETE":
+        try:
+            technician = Technician.objects.get(id=id)
+            technician.delete()
+            return JsonResponse(
+                technician,
+                encoder=TechnicianListEncoder,
+                safe=False
+            )
+        except Technician.DoesNotExist:
+            return(
+                JsonResponse({"message": "Technician does not exist"})
+            )
+    else:
         technician = Technician.objects.get(id=id)
-        technician.delete()
         return JsonResponse(
             technician,
             encoder=TechnicianListEncoder,
             safe=False
         )
-    except Technician.DoesNotExist:
-        return(
-            JsonResponse({"message": "Technician does not exist"})
+
+@require_http_methods(["GET", "POST"])
+def appointment_list(request):
+    if request.method == "GET":
+        appointments = Appointment.objects.all()
+        return JsonResponse(
+            {"appointments": appointments},
+            encoder=AppointmentListEncoder
+        )
+    else:
+        content = json.loads(request.body)
+        try:
+            technician_id = content["technician"]
+            technician = Technician.objects.get(id=technician_id)
+            content["technician"] = technician
+        except Technician.DoesNotExist:
+            return JsonResponse(
+                {"error": "Technician does not exist"},
+                status=404
+            )
+
+        appointment = Appointment.objects.create(**content)
+        return JsonResponse(
+            appointment,
+            encoder=AppointmentDetailEncoder,
+            safe=False
         )
